+++ lib/src/utils/utils.dart
@@ -2,6 +2,7 @@ library;
 
 import 'dart:async';
 import 'dart:convert';
+import 'dart:io';
 import 'dart:math';
 import 'dart:typed_data';
 import 'dart:developer' as dev;
@@ -18,3 +19,4 @@ part 'abi_coder.dart';
 part 'crypto.dart';
 part 'uint256.dart';
 part 'logger.dart';
+part '1271.dart';